Replaced old PoolCardTRI with new PoolCardTRIV2; Fixed AUR -> AURORA
