+++ src/components/earn/PoolCardTri.tsx
@@ -1,237 +1,162 @@
 import React from 'react'
 import { AutoColumn } from '../Column'
-import { RowBetween } from '../Row'
 import styled from 'styled-components'
-import { TYPE, StyledInternalLink } from '../../theme'
+import { TYPE } from '../../theme'
 import DoubleCurrencyLogo from '../DoubleLogo'
-import { CETH, Token, TokenAmount } from '@trisolaris/sdk'
+import { Token, TokenAmount } from '@trisolaris/sdk'
 import { ButtonPrimary } from '../Button'
-import { AutoRow } from '../Row'
+import { AutoRow, RowBetween } from '../Row'
 import { ChefVersions } from '../../state/stake/stake-constants'
-import { useColor } from '../../hooks/useColor'
+import { useColorForToken } from '../../hooks/useColor'
 import { currencyId } from '../../utils/currencyId'
-import { Break, CardNoise, CardBGImage } from './styled'
 import { unwrappedToken } from '../../utils/wrappedCurrency'
-import { PNG } from '../../constants'
 import { useTranslation } from 'react-i18next'
+import Card from '../Card'
+import { useHistory } from 'react-router-dom'
+import { addCommasToNumber } from '../../utils'
+import getTokenPairRenderOrder from '../../utils/getTokenPairRenderOrder'
+import { darken, lighten } from 'polished'
 
-const StatContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  flex-direction: column;
-  gap: 12px;
-  margin-bottom: 1rem;
-  margin-right: 1rem;
-  margin-left: 1rem;
-  ${({ theme }) => theme.mediaWidth.upToSmall`
-   display: none;
- `};
-`
-
-const AprContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  flex-direction: column;
-  gap: 12px;
-  margin-bottom: 1rem;
-  margin-right: 1rem;
-  margin-left: 1rem;
-`
-
-const Wrapper = styled(AutoColumn)<{ showBackground: boolean; bgColor1: any; bgColor2?: any }>`
-  border-radius: 12px;
-  width: 100%;
-  overflow: hidden;
-  position: relative;
-  opacity: 1;
-  background: ${({ theme, bgColor1, bgColor2, showBackground }) =>
-    bgColor2 != null
-      ? `radial-gradient(91.85% 100% at 1.84% 0%, ${bgColor1} 30%, ${bgColor2} 70%, ${
-          showBackground ? theme.black : theme.bg5
-        } 100%)`
-      : `radial-gradient(91.85% 100% at 1.84% 0%, ${bgColor1} 0%, ${showBackground ? theme.black : theme.bg5} 100%) `};
-  background-color: grey;
-  color: ${({ theme, showBackground }) => (showBackground ? theme.white : theme.text1)} !important;
-
-  ${({ showBackground }) =>
-    showBackground &&
-    `  box-shadow: 0px 0px 1px rgba(0, 0, 0, 0.01), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04),
-     0px 24px 32px rgba(0, 0, 0, 0.01);`}
-`
-
-const TopSection = styled.div`
+const Wrapper = styled(Card) < { bgColor1: string | null, bgColor2?: string | null, isDoubleRewards: boolean }>`
+  border: ${({ isDoubleRewards, theme }) =>
+        isDoubleRewards 
+            ? `1px solid ${theme.primary1}` 
+            : `1px solid ${theme.bg3};`
+    }
+  border-radius: 10px;
   display: grid;
-  grid-template-columns: 48px 1fr auto;
-  grid-gap: 0px;
-  align-items: center;
-  padding: 1rem;
-  z-index: 1;
-  ${({ theme }) => theme.mediaWidth.upToSmall`
-     grid-template-columns: 48px 1fr auto;
-   `};
+  grid-template-columns: 1fr;
+  grid-template-rows: 1fr 1fr;
+  gap: 12px;
+  box-shadow: ${({ isDoubleRewards, theme }) =>
+        isDoubleRewards 
+            ? `0px 0px 8px 5px ${theme.primary1}` 
+            : `0 2px 8px 0 ${theme.bg3}`
+    }
+  position: relative;
 `
 
-const BottomSection = styled.div<{ showBackground: boolean }>`
-  padding: 12px 16px;
-  opacity: ${({ showBackground }) => (showBackground ? '1' : '0.4')};
-  border-radius: 0 0 12px 12px;
-  display: flex;
-  flex-direction: row;
-  align-items: baseline;
-  justify-content: space-between;
-  z-index: 1;
-`
+const BackgroundColor = styled.span< { bgColor1: string | null, bgColor2?: string | null }>`
+   background: ${({ theme, bgColor1, bgColor2 }) =>
+        `linear-gradient(90deg, ${bgColor1 ?? theme.blue1} 0%, ${bgColor2 ?? 'grey'} 90%);`
+    }
+   background-size: cover;
+   mix-blend-mode: overlay;
+   border-radius: 10px;
+   width: 100%;
+   height: 100%;
+   opacity: 0.5;
+   position: absolute;
+   top: 0;
+   left: 0;
+   user-select: none;
+ `
 
-const GREY_ICON_TOKENS = ['ETH', 'WETH', 'WBTC', 'WNEAR']
-
-/**
- * token0
- * token1
- * stakedAmount
- */
+const PairContainer = styled.div`
+   display: flex;
+   align-items: center;
+ `
 
 type Props = {
-  apr: number
-  apr2: number
-  chefVersion: ChefVersions
-  isPeriodFinished: boolean
-  stakedAmount: TokenAmount | null
-  token0: Token
-  token1: Token
-  totalStakedInUSD: number
-  version: number
-  doubleRewards: boolean
-  inStaging: boolean 
+    apr: number
+    apr2: number
+    doubleRewards: boolean
+    chefVersion: ChefVersions
+    inStaging: boolean
+    isPeriodFinished: boolean
+    stakedAmount: TokenAmount | null
+    token0: Token
+    token1: Token
+    totalStakedInUSD: number
+    version: number
 }
 
-export default function PoolCard({
-  apr,
-  apr2,
-  chefVersion,
-  isPeriodFinished,
-  stakedAmount,
-  token0,
-  token1,
-  totalStakedInUSD,
-  version,
-  doubleRewards,
-  inStaging
+const Button = styled(ButtonPrimary) < { isStaking: boolean }>`
+  background: ${({ isStaking, theme }) => isStaking ? theme.black : darken(0.15, theme.primary1)}
+
+  ${({ isStaking, theme }) => isStaking && `
+        &:focus, &:hover, &:active {
+            background-color: ${lighten(0.12, theme.black)};
+        }
+  `}
+`
+
+export default function PoolCardTRI({
+    apr,
+    apr2,
+    chefVersion,
+    doubleRewards,
+    inStaging,
+    isPeriodFinished,
+    stakedAmount,
+    token0: _token0,
+    token1: _token1,
+    totalStakedInUSD,
+    version,
 }: Props) {
-  const currency0 = unwrappedToken(token0)
-  const currency1 = unwrappedToken(token1)
+    const [token0, token1] = getTokenPairRenderOrder(_token0, _token1)
+    const currency0 = unwrappedToken(token0)
+    const currency1 = unwrappedToken(token1)
 
-  const { t } = useTranslation()
-  const isStaking = Boolean(stakedAmount?.greaterThan('0') ?? false)
+    const { t } = useTranslation()
+    const isStaking = Boolean(stakedAmount?.greaterThan('0') ?? false)
+    const history = useHistory();
+    const isDualRewards = chefVersion == 1
 
-  const token: Token =
-    currency0 === CETH || currency1 === CETH
-      ? currency0 === CETH
-        ? token1
-        : token0
-      : token0.equals(PNG[token0.chainId])
-      ? token1
-      : token0
+    // get the color of the token
+    const backgroundColor1 = useColorForToken(token0)
 
-  // get the color of the token
-  let backgroundColor1 = useColor(token)
-  const secondaryToken = token === token1 ? token0 : token1
-  let backgroundColor2 = useColor(secondaryToken)
+    // Only override `backgroundColor2` if it's a dual rewards pool
+    const backgroundColor2 = useColorForToken(token1, () => isDualRewards);
 
-  const totalStakedInUSDFriendly = totalStakedInUSD.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',')
-  const isDualRewards = chefVersion == 1
+    const totalStakedInUSDFriendly = addCommasToNumber(totalStakedInUSD.toString())
+    return (
+        <Wrapper bgColor1={backgroundColor1} bgColor2={backgroundColor2} isDoubleRewards={doubleRewards}>
+            <BackgroundColor bgColor1={backgroundColor1} bgColor2={backgroundColor2} />
+            <AutoRow justifyContent="space-between">
+                <PairContainer>
+                    <DoubleCurrencyLogo currency0={currency0} currency1={currency1} size={24} />
+                    <TYPE.main marginLeft="0.5rem">
+                        {currency0.symbol}-{currency1.symbol}
+                    </TYPE.main>
+                </PairContainer>
+                <Button
+                    disabled={(isStaking || !isPeriodFinished) === false}
+                    isStaking={isStaking}
+                    padding="8px"
+                    borderRadius="10px"
+                    maxWidth="80px"
+                    onClick={() => {
+                        history.push(`/tri/${currencyId(currency0)}/${currencyId(currency1)}/${version}`)
+                    }}
+                >
+                    {isStaking ? t('earn.manage') : t('earn.deposit')}
+                </Button>
+            </AutoRow>
 
-  // Colors are dynamically chosen based on token logos
-  // These tokens are mostly grey; Override color to blue
-
-  if (GREY_ICON_TOKENS.includes(token?.symbol ?? '')) {
-    backgroundColor1 = '#2172E5'
-  }
-
-  // Only override `backgroundColor2` if it's a dual rewards pool
-  if (isDualRewards && GREY_ICON_TOKENS.includes(secondaryToken?.symbol ?? '')) {
-    backgroundColor2 = '#2172E5'
-  }
-
-  return (
-    <Wrapper showBackground={isStaking} bgColor1={backgroundColor1} bgColor2={isDualRewards ? backgroundColor2 : null}>
-      <CardBGImage desaturate />
-      <CardNoise />
-
-      <TopSection>
-        <DoubleCurrencyLogo currency0={currency0} currency1={currency1} size={24} />
-        <AutoRow align="baseline">
-          <TYPE.white fontWeight={600} fontSize={24} style={{ marginLeft: '8px' }}>
-            {currency0.symbol}-{currency1.symbol}
-          </TYPE.white>
-          {isDualRewards && doubleRewards ? (
-            <TYPE.white fontWeight={600} fontSize={16} style={{ marginLeft: '8px' }}>
-              Dual Rewards
-            </TYPE.white>
-          ) : null}
-        </AutoRow>
-        {(isStaking || !isPeriodFinished) && (
-          <StyledInternalLink
-            to={`/tri/${currencyId(currency0)}/${currencyId(currency1)}/${version}`}
-            style={{ width: '100%' }}
-          >
-            <ButtonPrimary padding="8px" borderRadius="8px">
-              {isStaking ? t('earn.manage') : t('earn.deposit')}
-            </ButtonPrimary>
-          </StyledInternalLink>
-        )}
-      </TopSection>
-
-      <StatContainer>
-        <RowBetween>
-          <TYPE.white> {t('earn.totalStaked')}</TYPE.white>
-          <TYPE.white>{`$${totalStakedInUSDFriendly}`}</TYPE.white>
-        </RowBetween>
-      </StatContainer>
-      {isDualRewards && doubleRewards ? (
-        <AprContainer>
-          <RowBetween>
-            <TYPE.white>TRI APR</TYPE.white>
-            <TYPE.white>{`${apr}%`}</TYPE.white>
-          </RowBetween>
-          <RowBetween>
-            <TYPE.white>AURORA APR</TYPE.white>
-            <TYPE.white>{`${apr2}%`}</TYPE.white>
-          </RowBetween>
-        </AprContainer>
-      ) : (
-        <AprContainer>
-          <RowBetween>
-            <TYPE.white>TRI APR</TYPE.white>
-            <TYPE.white>{inStaging ? `Coming Soon`: `${apr}%`}</TYPE.white>
-          </RowBetween>
-        </AprContainer>
-      )}
-      <StatContainer>
-        {/*<RowBetween>
-          <TYPE.white> {t('earn.poolWeight')} </TYPE.white>
-          <TYPE.white>{`${stakingInfo.multiplier}X`}</TYPE.white>
-        </RowBetween>*/}
-      </StatContainer>
-
-      {isStaking && (
-        <>
-          <Break />
-          {/*<BottomSection showBackground={true}>
-            <TYPE.black color={'white'} fontWeight={500}>
-              <span>{t('earn.yourRate')}</span>
-            </TYPE.black>
-
-            <TYPE.black style={{ textAlign: 'right' }} color={'white'} fontWeight={500}>
-              <span role="img" aria-label="wizard-icon" style={{ marginRight: '0.5rem' }}>
-                ⚡
-              </span>
-              {`${stakingInfo.rewardRate
-                ?.multiply(`${60 * 60 * 24 * 7}`)
-                ?.toSignificant(4, { groupSeparator: ',' })} PNG / week`}
-            </TYPE.black>
-          </BottomSection>*/}
-        </>
-      )}
-    </Wrapper>
-  )
+            <RowBetween>
+                <AutoColumn>
+                    <TYPE.subHeader>
+                        {t('earn.totalStaked')}
+                    </TYPE.subHeader>
+                    <TYPE.main>
+                        {`$${totalStakedInUSDFriendly}`}
+                    </TYPE.main>
+                </AutoColumn>
+                <AutoColumn >
+                    <TYPE.subHeader textAlign="end">
+                        APR
+                    </TYPE.subHeader>
+                    <TYPE.main textAlign="end">
+                        {(isDualRewards && doubleRewards
+                            ? `${apr}% TRI + ${`${apr2}%`} AURORA`
+                            : inStaging
+                                ? `Coming Soon`
+                                : `${apr}%`
+                        )}
+                    </TYPE.main>
+                </AutoColumn>
+            </RowBetween>
+        </Wrapper>
+    )
 }
